+++ src/edu/ucsb/cs56/games/client_server/v2/games/ServerControllers/TwoPlayerGameController.java
@@ -1,8 +1,10 @@
-package edu.ucsb.cs56.games.client_server.v2.server.Controllers;
+package edu.ucsb.cs56.games.client_server.v2.games.ServerControllers;
 
 import java.util.ArrayList;
 
-//import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaServer;
+import edu.ucsb.cs56.games.client_server.v2.server.Controllers.JavaServer;
+import edu.ucsb.cs56.games.client_server.v2.server.Controllers.ClientNetworkController;
+import edu.ucsb.cs56.games.client_server.v2.server.Controllers.ChatController;
 
 /**
  * an abstract service classification for games which have two players
